+++ packages/dao-utils/src/components/proposals/ProposalCardGrid.tsx
@@ -1,7 +1,7 @@
 /* @ts-ignore */
 import * as React from 'react'
 import ProposalCard from './ProposalCard'
-import { useGovernorContext } from '../context'
+import { useGovernorContext } from '../../context'
 
 export default function ProposalCardGrid() {
   const { proposals } = useGovernorContext()