+++ apps/next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # nextjs-dapp
 
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @public-assembly/dao-utils@0.0.23
+
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @public-assembly/dao-utils@0.0.22
+
 ## 1.0.16
 
 ### Patch Changes