+++ packages/dao-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @public-assembly/dao-utils
 
+## 0.0.23
+
+### Patch Changes
+
+- Fixes errors in previous version where components were not being exported properly.
+
+## 0.0.22
+
+### Patch Changes
+
+- Refactors the TokenWinningBid component to use the newly added useBid hook.
+
 ## 0.0.21
 
 ### Patch Changes